make signalresolve to handle WaitB if semaphore is not down'ed
originally, I model them as separate steps, where the signal resolve
mark it no longer signaled and increase counter by 1, and hopefully some
time later a new state will use WaitB to consume the counter and
unblock the thread. This is not correct. If the counter is 0 and the
wait queue of semaphore is empty, a signal (up) is considered atomic (at
the same time) the downer is unblocked. Thus this should be in the same
step and model the state transition of two concurrent execution: one
signaler is popping from the queue and the other one setting the thread
local semaphore to be 0 and try to lock.